Move rule configs out of step
