+++ src/bootstrap/native.rs
@@ -33,6 +33,16 @@ use Build;
 use util;
 use build_helper::up_to_date;
 
+/ rules.build("llvm", "src/llvm")
+//      .host(true)
+//      .dep(move |s| {
+//          if s.target == build.build {
+//              Step::noop()
+//          } else {
+//              s.target(&build.build)
+//          }
+//      })
+//      .run(move |s| native::llvm(build, s.target));
 /// Compile LLVM for `target`.
 pub fn llvm(build: &Build, target: &str) {
     // If we're using a custom LLVM bail out here, but we can only use a
@@ -216,6 +226,8 @@ fn check_llvm_version(build: &Build, llvm_config: &Path) {
     panic!("\n\nbad LLVM version: {}, need >=3.5\n\n", version)
 }
 
+//rules.build("test-helpers", "src/rt/rust_test_helpers.c")
+//     .run(move |s| native::test_helpers(build, s.target));
 /// Compiles the `rust_test_helpers.c` library which we used in various
 /// `run-pass` test suites for ABI testing.
 pub fn test_helpers(build: &Build, target: &str) {
@@ -252,6 +264,9 @@ pub fn test_helpers(build: &Build, target: &str) {
 const OPENSSL_VERS: &'static str = "1.0.2k";
 const OPENSSL_SHA256: &'static str =
     "6b3977c61f2aedf0f96367dcfb5c6e578cf37e7b8d913b4ecb6643c3cb88d8c0";
+
+//rules.build("openssl", "path/to/nowhere")
+//     .run(move |s| native::openssl(build, s.target));
 
 pub fn openssl(build: &Build, target: &str) {
     let out = match build.openssl_dir(target) {
